Factor out prvCheckOptions loop into function
The main loop in prvCheckOptions has been moved into a separate
function, prvCheckOptions_outerLoop. This is in preparation for CBMC
proofs that check properties of the loop separately from the rest of the
function, to keep the proofs scalable and easy to understand. It also
makes the function easier to read.

The new function takes all required arguments and local variables of the
calling function by reference, i.e. as pointer-to-pointers. It returns a
value that signals to the caller whether to break or continue the loop.
This commit adds a portable definition for MSVC's __forceinline
directive, which the new function uses.

This commit does not change the indentation of the new function, to
ensure that the diff shows only relevant changes. Whitespace will be
fixed in a separate commit.